+++ codex/librarian/importer/link_comics.py
@@ -5,12 +5,12 @@ from pathlib import Path
 from django.db.models import Q
 
 from codex.librarian.importer.const import (
-    CLASS_CUSTOM_COVER_GROUP_MAP,
     COMIC_FK_FIELD_NAMES,
     DICT_MODEL_FIELD_NAME_CLASS_MAP,
     DICT_MODEL_REL_LINK_MAP,
     FOLDERS_FIELD,
     IMPRINT,
+    M2M_MDS,
     PARENT_FOLDER,
     PUBLISHER,
     SERIES,
@@ -18,10 +18,10 @@ from codex.librarian.importer.const import (
     STORY_ARCS_METADATA_KEY,
     VOLUME,
 )
-from codex.librarian.importer.status import ImportStatusTypes, status_notify
+from codex.librarian.importer.link_covers import LinkCoversImporter
+from codex.librarian.importer.status import ImportStatusTypes
 from codex.models import (
     Comic,
-    CustomCover,
     Folder,
     Imprint,
     Publisher,
@@ -29,10 +29,10 @@ from codex.models import (
     Volume,
 )
 from codex.models.named import Identifier
-from codex.threads import QueuedThread
+from codex.status import Status
 
 
-class LinkComicsMixin(QueuedThread):
+class LinkComicsImporter(LinkCoversImporter):
     """Link comics methods."""
 
     @staticmethod
@@ -42,13 +42,12 @@ class LinkComicsMixin(QueuedThread):
         return md.get(field_name, group_class.DEFAULT_NAME)
 
     @classmethod
-    def get_comic_fk_links(cls, md, library, path):
+    def get_comic_fk_links(cls, md, path):
         """Get links for all foreign keys for creating and updating."""
         publisher_name = cls._get_group_name(Publisher, md)
         imprint_name = cls._get_group_name(Imprint, md)
         series_name = cls._get_group_name(Series, md)
         volume_name = cls._get_group_name(Volume, md)
-        md["library"] = library
         md[PUBLISHER] = Publisher.objects.get(name=publisher_name)
         md[IMPRINT] = Imprint.objects.get(
             publisher__name=publisher_name, name=imprint_name
@@ -124,13 +123,15 @@ class LinkComicsMixin(QueuedThread):
         result = frozenset(pks)
         all_m2m_links[key][comic_pk] = result
 
-    def _link_comic_m2m_fields(self, m2m_mds):
+    def _link_comic_m2m_fields(self):
         """Get the complete m2m field data to create."""
         all_m2m_links = {}
-        comic_paths = frozenset(m2m_mds.keys())
+        comic_paths = frozenset(self.metadata.get(M2M_MDS, {}).keys())
+        if not comic_paths:
+            return all_m2m_links
         comics = Comic.objects.filter(path__in=comic_paths).values_list("pk", "path")
         for comic_pk, comic_path in comics:
-            md = m2m_mds[comic_path]
+            md = self.metadata[M2M_MDS][comic_path]
             link_data = (all_m2m_links, md, comic_pk)
             self._link_prepare_special_m2ms(
                 link_data, FOLDERS_FIELD, Folder, self._get_link_folders_filter
@@ -143,6 +144,7 @@ class LinkComicsMixin(QueuedThread):
                 self._link_prepare_special_m2ms(link_data, field_name, model, method)
 
             self._link_named_m2ms(all_m2m_links, comic_pk, md)
+        self.metadata.pop(M2M_MDS)
         return all_m2m_links
 
     def _query_relation_adjustments(
@@ -204,10 +206,9 @@ class LinkComicsMixin(QueuedThread):
         tms, all_del_pks = self._query_relation_adjustments(
             m2m_links, ThroughModel, through_field_id_name, status
         )
-        if status:
-            status.total = status.total or 0
-            status.total += len(tms) + len(all_del_pks)
-            self.status_controller.update(status)
+        status.total = status.total or 0
+        status.total += len(tms) + len(all_del_pks)
+        self.status_controller.update(status)
 
         update_fields = ("comic_id", through_field_id_name)
         ThroughModel.objects.bulk_create(
@@ -218,10 +219,9 @@ class LinkComicsMixin(QueuedThread):
         )
         if created_count := len(tms):
             self.log.info(f"Linked {created_count} new {field_name} to altered comics.")
-        if status:
-            status.complete = status.complete or 0
-            status.complete += created_count
-            self.status_controller.update(status)
+        status.complete = status.complete or 0
+        status.complete += created_count
+        self.status_controller.update(status)
 
         if del_count := len(all_del_pks):
             del_qs = ThroughModel.objects.filter(pk__in=all_del_pks)
@@ -229,19 +229,16 @@ class LinkComicsMixin(QueuedThread):
             self.log.info(
                 f"Deleted {del_count} stale {field_name} relations for altered comics.",
             )
-        if status:
-            status.complete = status.complete or 0
-            status.complete += del_count
-            self.status_controller.update(status)
+        status.complete = status.complete or 0
+        status.complete += del_count
+        self.status_controller.update(status)
         return created_count, del_count
 
-    @status_notify(status_type=ImportStatusTypes.LINK_M2M_FIELDS, updates=False)
-    def bulk_query_and_link_comic_m2m_fields(self, all_m2m_mds, status=None):
+    def bulk_query_and_link_comic_m2m_fields(self):
         """Combine query and bulk link into a batch."""
-        if status:
-            status.complete = 0
-            status.total = 0
-        all_m2m_links = self._link_comic_m2m_fields(all_m2m_mds)
+        status = Status(ImportStatusTypes.LINK_M2M_FIELDS)
+        self.status_controller.start(status)
+        all_m2m_links = self._link_comic_m2m_fields()
         created_total = 0
         del_total = 0
         for field_name, m2m_links in all_m2m_links.items():
@@ -258,57 +255,5 @@ class LinkComicsMixin(QueuedThread):
             self.log.info(f"Linked {created_total} comics to tags.")
         if del_total:
             self.log.info(f"Deleted {del_total} stale relations for altered comics.")
+        self.status_controller.finish(status)
         return created_total + del_total
-
-    def _link_custom_cover_prepare(self, cover, model_map):
-        """Prepare one cover in the model map for bulk update."""
-        if cover.library and cover.library.covers_only:
-            model = CLASS_CUSTOM_COVER_GROUP_MAP.inverse.get(cover.group)
-            if not model:
-                self.log.warning(f"Custom Cover model not found for {cover.path}")
-                return
-            group_filter = {"sort_name": cover.sort_name}
-        else:
-            model = Folder
-            path = str(Path(cover.path).parent)
-            group_filter = {"path": path}
-        qs = model.objects.filter(**group_filter).exclude(custom_cover=cover)
-        if not qs.exists():
-            return
-
-        if model not in model_map:
-            model_map[model] = []
-
-        for obj in qs.iterator():
-            obj.custom_cover = cover
-            model_map[model].append(obj)
-
-    def _link_custom_cover_group(self, model, objs, status):
-        """Bulk link a group to it's custom covers."""
-        count = 0
-        if not objs:
-            return count
-        model.objects.bulk_update(objs, ["custom_cover"])
-        count += len(objs)
-        self.log.info(f"Linked {count} custom covers to {model.__name__}s")
-        if status:
-            status.complete = status.complete or 0
-            status.complete += count
-        return count
-
-    @status_notify(status_type=ImportStatusTypes.COVERS_LINK)
-    def link_custom_covers(self, link_cover_pks, status=None):
-        """Link Custom Covers to Groups."""
-        # Aggregate objs to update for each group model.
-        model_map = {}
-        covers = CustomCover.objects.filter(pk__in=link_cover_pks).only(
-            "library", "path"
-        )
-        for cover in covers:
-            self._link_custom_cover_prepare(cover, model_map)
-
-        # Bulk update each model type
-        total_count = 0
-        for model, objs in model_map.items():
-            total_count += self._link_custom_cover_group(model, objs, status)
-        return total_count