+++ codex/librarian/importer/status.py
@@ -1,12 +1,6 @@
 """Librarian Status for imports."""
 
-from collections.abc import Callable
-from enum import Enum
-from time import time
-
 from django.db.models import Choices
-
-from codex.status import Status
 
 
 class ImportStatusTypes(Choices):
@@ -31,36 +25,4 @@ class ImportStatusTypes(Choices):
     COVERS_CREATED = "ICC"
     COVERS_DELETED = "ICD"
     COVERS_LINK = "ICL"
-
-
-def status_notify(status_type: str | Enum = "", updates=True):
-    """Wrap a function with status changes."""
-    # https://stackoverflow.com/questions/5929107/decorators-with-parameters
-
-    def decorator(func) -> Callable[..., int]:
-        def wrapper(self, data, *args, status=None, **kwargs) -> int:
-            """Run a function bracketed by status changes."""
-            num_elements = len(data)
-            if not num_elements:
-                return 0
-
-            if status:
-                finish = False
-            else:
-                complete = 0 if updates else None
-                status = Status(status_type, complete, num_elements, time())
-                self.status_controller.start(status)
-                finish = True
-
-            kwargs["status"] = status
-            try:
-                count = func(self, data, *args, **kwargs)
-            finally:
-                if finish:
-                    self.status_controller.finish(status)
-
-            return count
-
-        return wrapper
-
-    return decorator
+    GROUP_UPDATE = "IGU"