+++ codex/librarian/importer/tasks.py
@@ -1,6 +1,7 @@
 """DB Import Tasks."""
 
-from dataclasses import dataclass
+from collections.abc import Mapping
+from dataclasses import dataclass, field
 
 
 @dataclass
@@ -14,20 +15,20 @@ class ImportDBDiffTask(ImportTask):
 
     library_id: int
 
-    dirs_moved: dict[str, str]
-    dirs_modified: frozenset[str]
-    # dirs_created: frozenset[str]
-    dirs_deleted: frozenset[str]
+    dirs_moved: Mapping[str, str] = field(default_factory=dict)
+    dirs_modified: frozenset[str] = frozenset()
+    # dirs_created: frozenset[str] | None = frozenset()
+    dirs_deleted: frozenset[str] = frozenset()
 
-    files_moved: dict[str, str]
-    files_modified: frozenset[str]
-    files_created: frozenset[str]
-    files_deleted: frozenset[str]
+    files_moved: Mapping[str, str] = field(default_factory=dict)
+    files_modified: frozenset[str] = frozenset()
+    files_created: frozenset[str] = frozenset()
+    files_deleted: frozenset[str] = frozenset()
 
-    covers_moved: dict[str, str]
-    covers_modified: frozenset[str]
-    covers_created: frozenset[str]
-    covers_deleted: frozenset[str]
+    covers_moved: Mapping[str, str] = field(default_factory=dict)
+    covers_modified: frozenset[str] = frozenset()
+    covers_created: frozenset[str] = frozenset()
+    covers_deleted: frozenset[str] = frozenset()
 
     force_import_metadata: bool = False